[fix] session 및 uri 연결 (#57)
인증로직 필요한 Controller에 인증로직 추가

personal-header & company-header
ㄴ 주소 변경 시 css 깨지는 오류 잡음 -> css 절대경로로 변경

인증 필요한 주소는 /s/~ , 인증필요 & data return 주소는 /s/api/~ 변경

UserController
ㄴ /s/company/user/update-form
ㄴ /s/company/user/update
ㄴ /s/personal/user/update-form
ㄴ /s/personal/user/update
ㄴ /s/company/main

ResumeController
(기업&개인이 공통으로 사용하는 커뮤니티 제외하고는 role이 구분되므로 /s/personal/~ 로 통일함)
ㄴ /s/personal/resume
ㄴ /s/personal/resume/{id}
ㄴ /s/personal/resume/{id}/delete
ㄴ redirect:/s/personal/resume
ㄴ /s/personal/resume/save-form
ㄴ /s/api/personal/resume/save

JobPostingController
(personal과 company가 같이 있어서 기업관련 controller 메서드는 /s/company/~로 통일함)
ㄴ /s/company/jobposting/list
ㄴ /s/company/jobposting/{id}
ㄴ /s/company/jobposting/save-form
ㄴ /s/company/jobposting/save
ㄴ /s/company/jobposting/{id}/update-form
ㄴ /s/company/jobposting/{id}/update
ㄴ /s/company/jobposting/{id}/delete
ㄴ redirect:/s/company/jobposting/list
ㄴ "redirect:/s/company/jobposting/" + id

JobPostingBookmarkController
ㄴ /s/api/personal/bookmark
ㄴ /s/api/personal/bookmark/{id}

CompanyInfoController
ㄴ /s/company/info/{id}
ㄴ /s/company/info/save-form
ㄴ /s/company/info/save
ㄴ /s/company/info/{id}/update-form
ㄴ /s/company/info/{id}/update
ㄴ /s/company/info/{id}/delete
ㄴ "redirect:/s/company/info/" + companyInfo.getId()
ㄴ "redirect:/s/company/info/" + id;
ㄴ redirect:/s/company/main

ApplicationController
ㄴ /s/personal/mypage/bookmark
ㄴ /s/personal/mypage/application
ㄴ /s/company/applicant/list
ㄴ /s/company/applicant/{id}
ㄴ /s/api/company/applicant/{id}/pass
ㄴ /s/personal/jobposting/{id}/apply-form

ApplicationBookmarkController
ㄴ /s/api/company/bookmark
ㄴ /s/api/company/bookmark/{id}

LikeController
ㄴ /s/api/like
ㄴ /s/api/like/{id}

ReplyController
ㄴ /s/reply/save
ㄴ /s/reply/{id}/delete

BoardController
ㄴ /s/community/save-form
ㄴ /s/community/save
ㄴ /s/community/{id}/update-form
ㄴ /s/community/{id}/update
ㄴ /s/community/{id}/delete

mustache 파일 주소 연결 수정
ㄴ 이력서 상세보기 내 수정하기 버튼 주소 변경 -> /s/personal/resume/{{model.id}}/update-form
ㄴ 지원하기 주소 변경 -> /s/personal/apply

mustache 변경사항
ㄴ 댓글 수정버튼 삭제
ㄴ /s/company/jobposting/list(포지션)에서 로그인한 기업유저의 채용 공고만 확인되도록 수정
ㄴ company/applicant/list(지원자관리)에서 스크랩(별) 버튼 누르면 지원자상세페이지로 이동되는 것을 스크랩버튼만 동작하도록 수정

code 변경사항
ㄴ /s/company/jobposting/list(포지션) 클릭 시 해당 기업 user의 채용포지션이 아닌 전체 공고가 뜨는 문제 발생 -> 로그인 한 기업 user의 채용 포지션만 확인되도록 JobPostingController, JobPostingService, JobPostingRepository 수정

WebMvcConfig, LoginInterceptor 내 코드 주석처리함

+++ src/main/java/com/example/leapit/resume/ResumeController.java
@@ -18,36 +18,36 @@ public class ResumeController {
     private final ResumeService resumeService;
     private final HttpSession session;
 
-    @GetMapping("/resume")
+    @GetMapping("/s/personal/resume")
     public String list(HttpServletRequest request) {
-//        User sessionUser = (User) session.getAttribute("sessionUser");
-//        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
+       User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
         Integer sessionUserId = 1;
         List<Resume> resumeList = resumeService.list(1); // TODO : sessionUser.getId() 인수 추가
         request.setAttribute("models", resumeList);
         return "personal/resume/list";
     }
 
-    @GetMapping("/resume/{id}")
+    @GetMapping("/s/personal/resume/{id}")
     public String detail(@PathVariable("id") int id, HttpServletRequest request) {
-//        User sessionUser = (User) session.getAttribute("sessionUser");
-//        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
+      User sessionUser = (User) session.getAttribute("sessionUser");
+      if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
 
         ResumeResponse.DetailDTO detailDTO = resumeService.detail(id); // TODO : sessionUser.getId() 인수 추가
         request.setAttribute("model", detailDTO);
         return "personal/resume/detail";
     }
 
-    @PostMapping("/resume/{id}/delete")
+    @PostMapping("/s/personal/resume/{id}/delete")
     public String delete(@PathVariable("id") int id) {
-//        User sessionUser = (User) session.getAttribute("sessionUser");
-//        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
+       User sessionUser = (User) session.getAttribute("sessionUser");
+       if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
 
         resumeService.delete(id); // TODO : sessionUser.getId() 인수 추가
-        return "redirect:/resume";
+        return "redirect:/s/personal/resume";
     }
 
-    @GetMapping("/resume/save-form")
+    @GetMapping("/s/personal/resume/save-form")
     public String saveForm(HttpServletRequest request) {
         User sessionUser = (User) session.getAttribute("sessionUser");
         if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
@@ -57,7 +57,7 @@ public class ResumeController {
         return "personal/resume/save-form";
     }
 
-    @PostMapping("/resume/save")
+    @PostMapping("/s/api/personal/resume/save")
     @ResponseBody
     public Resp<?> save(@RequestBody ResumeRequest.SaveDTO saveDTO) {
         User sessionUser = (User) session.getAttribute("sessionUser");
